+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - craftcms 4.2.0.2
 - npm packages
 
+### Changed
+- Changed the logic how colors are defined. tailwind.settings.colors.js 
+
 
 ## 4.0.6 - 2022.07.26